+++ stag_lib/graph.h
@@ -57,6 +57,9 @@ typedef Eigen::Triplet<double, stag_int> EdgeTriplet;
 // Redefine the eigen index type to be the same as stag_int
 #undef EIGEN_DEFAULT_DENSE_INDEX_TYPE
 #define EIGEN_DEFAULT_DENSE_INDEX_TYPE stag_int
+
+// Define a small epsilon
+#define EPSILON 0.0000000001
 /**
  * \endcond
  */
@@ -172,11 +175,19 @@ namespace stag {
    * Vertices of the graph are always referred to by their unique integer index.
    * This index corresponds to the position of the vertex in the stored adjacency
    * matrix of the graph.
+   *
+   * This class supports graphs with positive edge weights. Self-loops are
+   * permitted.
    */
   class Graph : public LocalGraph {
     public:
       /**
        * Create a graph from an Eigen matrix.
+       *
+       * The provided matrix should correspond either to the
+       * adjacency matrix or Laplacian matrix of the graph. STAG will
+       * automatically detect whether the provided matrix is an adjacency matrix
+       * or a Laplacian matrix.
        *
        * \par Example
        *
@@ -206,13 +217,14 @@ namespace stag {
        * }
        * \endcode
        *
-       * The provided adjacency matrix must be symmetric.
+       * The provided matrix must be symmetric, and may include
+       * self-loops.
        *
-       * @param adjacency_matrix the sparse eigen matrix representing the adjacency matrix
-       *               of the graph.
-       * @throws domain_error if the adjacency matrix is not symmetric
+       * @param matrix the sparse eigen matrix representing the adjacency matrix
+       *               or Laplacian matrix of the graph.
+       * @throws domain_error if the provided matrix is not symmetric
        */
-      explicit Graph(const SprsMat& adjacency_matrix);
+      explicit Graph(const SprsMat& matrix);
 
       /**
        * Create a graph from raw arrays describing a CSC sparse matrix.
@@ -380,6 +392,11 @@ namespace stag {
        stag_int number_of_edges() const;
 
        /**
+        * Returns a boolean indicating whether this graph contains self loops.
+        */
+       bool has_self_loops() const;
+
+       /**
         * Construct and return a subgraph of this graph.
         *
         * Note that the vertex indices will be changed in the subgraph.
@@ -487,6 +504,9 @@ namespace stag {
       // matrix, stored in a sparse format. The adj_init_ variable is used to
       // indicate whether the matrix has been initialised yet.
       SprsMat adjacency_matrix_;
+
+      // Whether the graph has self loops
+      bool has_self_loops_;
 
       // The laplacian matrix of the graph. The lap_init_ variable is used to
       // indicate whether the matrix has been initialised yet.