+++ stag_lib/CHANGELOG.md
@@ -12,6 +12,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - [Issue #93](https://github.com/staglibrary/stag/issues/93): Add methods to compute mutual information between clusters
 - [Issue #98](https://github.com/staglibrary/stag/issues/98): Add method to construct the graph union
 - [Issue #22](https://github.com/staglibrary/stag/issues/22): Add Cheeger cut method
+- [Issue #107](https://github.com/staglibrary/stag/issues/107): Add support for self-loops in graphs
+- [Issue #108](https://github.com/staglibrary/stag/issues/108): Allow graphs to be initialised with Laplacian matrix
 
 ### Fixed
 - [Issue #87](https://github.com/staglibrary/stag/issues/87): Occasional bug with sorting edgelist file